Update project structure - Add containers folder
