+++ src/containers/EditMeasurePage.js
@@ -1,10 +1,11 @@
-import React, { useState } from 'react';
+/* eslint-disable react/jsx-props-no-spreading */
+import { useHistory, useParams } from 'react-router-dom';
+import { useState } from 'react';
 import { connect } from 'react-redux';
-import { useHistory } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { startAddMeasure } from '../actions/measures';
+import { startEditMeasure } from '../actions/measures';
 import Header from './Header';
-import Loader from './Loader';
+import Loader from '../components/Loader';
 import MeasureForm from './MeasureForm';
 import {
   container as formContainer,
@@ -12,14 +13,15 @@ import {
 } from '../stylesheet/Form.module.css';
 import { container, fullHeight } from '../stylesheet/CommonPage.module.css';
 
-const AddMeasurePage = ({ startAddMeasure }) => {
+const EditMeasurePage = ({ startEditMeasure, measure }) => {
+  const { id } = useParams();
   const history = useHistory();
 
   const [loading, setLoading] = useState(false);
 
-  const handleSubmit = measure => {
+  const handleSubmit = updates => {
     setLoading(true);
-    startAddMeasure(measure).then(({ error }) => {
+    startEditMeasure(Number(id), updates).then(({ error }) => {
       if (error) {
         setLoading(false);
       } else {
@@ -35,11 +37,13 @@ const AddMeasurePage = ({ startAddMeasure }) => {
           <Loader height={fullHeight} />
         ) : (
           <div>
-            <Header title="Add Measure" back />
+            <Header title="Edit Measure" back />
             <div className={`${container}`}>
               <div className={`${formContainer} ${mNone}`}>
                 <MeasureForm
+                  {...measure}
                   onSubmit={handleSubmit}
+                  update
                 />
               </div>
             </div>
@@ -50,12 +54,17 @@ const AddMeasurePage = ({ startAddMeasure }) => {
   );
 };
 
-AddMeasurePage.propTypes = {
-  startAddMeasure: PropTypes.func.isRequired,
+EditMeasurePage.propTypes = {
+  startEditMeasure: PropTypes.func.isRequired,
+  measure: PropTypes.objectOf(PropTypes.any).isRequired,
 };
 
 const mapDispatchToProps = {
-  startAddMeasure,
+  startEditMeasure,
 };
 
-export default connect(null, mapDispatchToProps)(AddMeasurePage);
+const mapStateToProps = ({ measures }, props) => ({
+  measure: measures.find(measure => measure.id === Number(props.match.params.id)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(EditMeasurePage);
