+++ src/containers/ScorePage.js
@@ -4,8 +4,8 @@ import moment from 'moment';
 import { Link } from 'react-router-dom';
 import { repGoal } from '../usefull_vars';
 import Header from './Header';
-import MeasurementDisplay from './MeasurementDisplay';
-import ProgressCircle from './ProgressCircle';
+import MeasurementDisplay from '../components/MeasurementDisplay';
+import ProgressCircle from '../components/ProgressCircle';
 import { findMeasurementByDate, getScore } from '../helpers/measurements';
 import next from '../assets/images/next.png';
 import prev from '../assets/images/prev.png';
