+++ src/containers/TrackPage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import Header from './Header';
-import MeasurementScore from './MeasurementScore';
+import MeasurementScore from '../components/MeasurementScore';
 import {
   container,
 } from '../stylesheet/CommonPage.module.css';
